Move export api content
